Creates Alternative Key Delimiter for Secrets Manager Keys
Dynamic references require the use of colon (:) as delimiter for
segments of the reference, so the colon-combined keys we have been
using to flatten secret configuration into key–value pairs abolutely
will not work. Fortunately, we have some prior art in the environment
variable configuration, which is another place where colons are
forbidden. The method of normalization for those has been applied to
the flattened keys here.

This is a backwards-compatible change. Colon continues to work, but
dunder is now _also_ accepted.